优化系统管理-用户管理信息

+++ 02-security-system/src/main/java/cn/wickson/security/system/model/entity/SystemUser.java
@@ -27,7 +27,7 @@ import java.time.LocalDateTime;
 public class SystemUser extends BaseDO {
 
     @TableId
-    private BigInteger id;
+    private Long id;
 
     /**
      * 用户名
@@ -65,7 +65,7 @@ public class SystemUser extends BaseDO {
     private String mobile;
 
     /**
-     * 用户状态((1:正常;0:禁用))
+     * 用户状态(1:正常;0:禁用)
      */
     private Integer status;
